lib: split into several modules
